requirements.txt
pyupbit==0.2.14
stockstats==0.3.2
python-telegram-bot==13.5
